+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to the Audio Player project will be documented in this file.
 
 ## 3.8.0 - 2026-03-18
 ### Fixed 
-- replaced deprecated QueryBuilder `execute()` calls with `executeQuery()`/`executeStatement()
+- NC33: replaced deprecated QueryBuilder `execute()` calls with `executeQuery()`/`executeStatement()
+- NC33: replaced deprecated controller statement `fetchAll()` calls with `fetchAllAssociative()`
 
 ## 3.7.0 - 2025-11-24
 ### Changed